Relax to proper equilibrium state
Relaxation now takes the vector potential into account (the equilibrium state with a constant vector potential is a different one)
This does not (yet) work on the HoneyComb lattice (may related to the strange BZ and complex phase of gamma?), but on PiFlux everything seems fine.
Next update:
Diagonal vs. off diagonal relaxation